Fix np float dtype

+++ eye_tracking/gaze/head_pose_estimation/head_pose_normalizer.py
@@ -59,5 +59,5 @@ class HeadPoseNormalizer:
                 [0, 1, 0],
                 [0, 0, self.normalized_distance / distance],
             ],
-            dtype=np.float,
+            dtype=np.float32,
         )
